tts-livekit-plugin/tests/requirements.txt
# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
aiohttp>=3.9.0

# Plugin dependencies (for testing)
livekit-agents>=0.8.0
